moves sessiontimeout into LoginAlert
