--- src/platform/user/authentication/components/SessionTimeoutAlert.jsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
-import { VaAlert } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-import { renderMaintenanceWindow, renderDowntimeBanner } from '../downtime';
-
-export default function SessionTimeoutAlert() {
-  const { statuses = [], maintenanceWindows = [] } = useSelector(
-    state => state.externalServiceStatuses,
-  );
-  const noDowntime =
-    !renderDowntimeBanner(statuses) &&
-    !renderMaintenanceWindow(maintenanceWindows);
-  const isSessionExpired = window.location.href.includes(
-    '/?next=loginModal&status=sessionExpired',
-  );
-  const displaySessionTimeout = isSessionExpired && noDowntime;
-  return displaySessionTimeout ? (
-    <VaAlert closeable={false} showIcon uswds>
-      <p slot="headline">Your session timed out. Sign in again to continue.</p>
-    </VaAlert>
-  ) : null;
-}
